\name{as.data.frame-methods}
\docType{methods}
\alias{as.data.frame-methods}
\alias{as.data.frame,ANY-method}
\alias{as.data.frame,desOpt-method}
\alias{as.data.frame,facDesign-method}
\alias{as.data.frame,gageRR-method}
\title{Methods for Function as.data.frame in Package base}
\description{Methods for function \code{as.data.frame} in Package base}



\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
ANY
}

\item{\code{signature(x = "desOpt")}}{
object of class desOpt
}

\item{\code{signature(x = "facDesign")}}{
object of class facDesign
}

\item{\code{signature(x = "gageRR")}}{
object of class gageRR
}}}

\keyword{methods}

