\name{as.data.frame.gageRR}
\alias{as.data.frame.gageRR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coerce to a data.frame
}
\description{
S3 generic for gageRR class
}

\usage{
\method{as.data.frame}{gageRR}(x, row.names = NULL, optional = FALSE, ...)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class gageRR
}
  \item{row.names}{
row.names
}
  \item{optional}{
optional
}
  \item{\dots}{
...
}
}
\value{
data.frame}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}


