\name{as.data.frame.mixDesign}
\alias{as.data.frame.mixDesign}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coerce the object of class mixDesign to a data.frame
}
\description{
S3 generic for mixDesign class
}

\usage{
\method{as.data.frame}{mixDesign}(x, row.names = NULL, optional = FALSE, ...)
}


%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class mixDesign
}
  \item{row.names}{
row.names
}
  \item{optional}{
optional
}
  \item{\dots}{
...
}
}
\value{
data.frame}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}


