\name{as.data.frame.MSALinearity}
\alias{as.data.frame.MSALinearity}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coerce to a data.frame
}
\description{
S3 generic for MSALinearity class
}

\usage{
\method{as.data.frame}{MSALinearity}(x, row.names = NULL, optional = FALSE, ...)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class MSALinearity
}
  \item{row.names}{
row.names
}
  \item{optional}{
optional
}
  \item{\dots}{
...
}
}
\value{
data.frame}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}


