\name{averagePlot-methods}
\docType{methods}
\alias{averagePlot-methods}
\alias{averagePlot,gageRR-method}
\title{Methods for Function averagePlot}
\description{
Methods for function \code{averagePlot}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "gageRR")}}{
function to create an average plot for an object of class \code{\link{gageRR}}.
}

}}
\seealso{
\url{http://www.r-qualitytools.org}
}
\keyword{methods}
\keyword{design}
