\name{code2real}
\alias{code2real}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coding
}
\description{
calculates the real value of a coded value
}
\usage{
code2real(low, high, codedValue)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{low}{
numeric
}
  \item{high}{
numeric
}
  \item{codedValue}{
numeric
}
}
\author{Thomas Roth \email{thomas.roth@tu-berlin.de}}
\examples{
code2real(160, 200, 0)

}