\name{compPlot-methods}
\docType{methods}
\alias{compPlot-methods}
\alias{compPlot,gageRR-method}
\title{Methods for Function compPlot}
\description{
Methods for function \code{compPlot}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "gageRR")}}{
function to create an comparison plot for an object of class \code{\link{gageRR}}.
}

}}
\seealso{
\url{http://www.r-qualitytools.org}
}
\keyword{methods}
\keyword{design}
