\name{confounds}
\alias{confounds}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Confounded Effects
}
\description{
Function to display confounded effects of a fractional factorial design in a human readable way.
}
\usage{
confounds(x, depth = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class facDesign
}
  \item{depth}{
numeric - up to depth-way confounded interactions are printed
}
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}

