\name{desirability-class}
\keyword{Multiple Response Optimization}
\Rdversion{1.1}
\docType{class}
\alias{desirability-class}


\title{Class "desirability"}
\description{
desirability Class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("desirability", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{response}:}{Object of class \code{"character"} ~~ }
    \item{\code{low}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{high}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{target}:}{Object of class \code{"ANY"} ~~ }
    \item{\code{scale}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{importance}:}{Object of class \code{"numeric"} ~~ }
  }
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\examples{
showClass("desirability")
}
\keyword{classes}
