\name{desOpt-class}
\Rdversion{1.1}
\docType{class}
\alias{desOpt-class}

\title{Class "desOpt"}
\description{
desOpt Class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("desOpt", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{facCoded}:}{Object of class \code{"list"} ~~ }
    \item{\code{facReal}:}{Object of class \code{"list"} ~~ }
    \item{\code{responses}:}{Object of class \code{"list"} ~~ }
    \item{\code{desirabilities}:}{Object of class \code{"list"} ~~ }
    \item{\code{overall}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{all}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{fdo}:}{Object of class \code{"facDesign"} ~~ }
  }
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}

\examples{
showClass("desOpt")
}
\keyword{classes}
