\name{distrCollection-class}
\Rdversion{1.1}
\docType{class}
\alias{distrCollection-class}

\title{Class "distrCollection"}
\description{
distrCollection Class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("distrCollection", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{distr}:}{Object of class \code{"list"} ~~ }
  }
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\examples{
showClass("distrCollection")
}
\keyword{classes}
