\name{doeFactor-class}
\Rdversion{1.1}
\docType{class}
\alias{doeFactor-class}

\title{Class "doeFactor"}
\description{
doeFactor Class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("doeFactor", ...)}.
}

\section{Slots}{
  \describe{
    \item{\code{low}:}{Object of class \code{"ANY"} ~~ }
    \item{\code{high}:}{Object of class \code{"ANY"} ~~ }
    \item{\code{name}:}{Object of class \code{"character"} ~~ }
    \item{\code{unit}:}{Object of class \code{"character"} ~~ }
    \item{\code{type}:}{Object of class \code{"character"} ~~ }
  }
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\examples{
showClass("doeFactor")
}
\keyword{classes}
