\name{effectPlot-methods}
\docType{methods}
\alias{effectPlot-methods}
\alias{effectPlot,facDesign-method}
\alias{effectPlot,taguchiDesign-method}
\title{Methods for Function effectPlot}
\description{
Methods for function \code{effectPlot}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "facDesign")}}{
function to create an effect plot for an object of class \code{\link{facDesign}}.
}

\item{\code{signature(object = "taguchiDesign")}}{
function to create an effect plot for an object of class \code{\link{taguchiDesign}}.
}
}}
\seealso{
\url{http://www.user.tu-berlin.de/kalicete/qualityTools}
}
\keyword{methods}
\keyword{design}
