\name{errorPlot-methods}
\docType{methods}
\alias{errorPlot-methods}
\alias{errorPlot,gageRR-method}
\title{Methods for Function errorPlot}
\description{
Methods for function \code{errorPlot}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "gageRR")}}{
function to create an error plot for an object of class \code{\link{gageRR}}.
}

}}
\seealso{
\url{http://www.r-qualitytools.org}
}
\keyword{methods}
\keyword{design}
