\name{fracChoose}
\Rdversion{1.1}
\alias{fracChoose}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Choosing a fractional or full factorial design from a table
}
\description{
Designs displayed are the classic minimum abberation designs. Choosing a design is done by clicking with the mouse into the appropriate field.

}
\usage{
fracChoose()
}

\value{
an object of class facDesign
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\note{
Statistics for Experimenters (2005) Box, Hunter, Hunter p.272, table 6.22
}

\seealso{
\code{\link{facDesign}}
\code{\link{fracDesign}}
\code{\link{rsmChoose}}
\code{\link{rsmDesign}}
}
