\name{nrow-methods}
\docType{methods}
\alias{nrow}
\alias{nrow-methods}
\alias{nrow,facDesign-method}
\title{Get method}
\description{Get the number of rows for a facDesign object.}

\usage{
\S4method{nrow}{facDesign}(x)
}
\arguments{
\item{x}{a \code{\link{facDesign}} object}
}

\value{numeric - number of rows}

\section{Methods}{
\describe{
\item{\code{signature(x = "facDesign")}}{
Get the number of rows for a \code{\link{facDesign}} object.
}
}}
\author{Thomas Roth \email{thomas.roth@tu-berlin.de}}
\seealso{
\code{\link{nrow}}\cr
\url{http://www.r-qualitytools.org}
}
