\name{pbFactor-class}
\Rdversion{1.1}
\docType{class}
\alias{pbFactor-class}

\title{Class "pbFactor"}
\description{
pbFactor Class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("pbFactor", ...)}.
}


\section{Slots}{
  \describe{
    \item{\code{values}:}{Object of class \code{"ANY"} ~~ }
    \item{\code{name}:}{Object of class \code{"character"} ~~ }
    \item{\code{unit}:}{Object of class \code{"character"} ~~ }
    \item{\code{type}:}{Object of class \code{"character"} ~~ }
  }
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
Etienne Stockhausen \email{stocdarf@mailbox.tu-berlin.de}
}
\examples{
showClass("pbFactor")
}
\keyword{classes}
