\name{print.adtest}
\alias{print.adtest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Test Statistics
}
\description{
Generic S3 function for objects of class adtest
}
\usage{
print.adtest(x, digits = 4, quote = TRUE, prefix = "", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class adtest
}
  \item{digits}{
numeric
}
  \item{quote}{
TRUE/FALSE
}
  \item{prefix}{
character
}
  \item{\dots}{
...
}
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}


