\name{print.invisible}
\alias{print.invisible}
\title{
Print Function
}
\description{
Generic S3 function for objects of class invisible
}
\usage{
print.invisible(x, ...)
}
\arguments{
  \item{x}{
an object of class invisible}
  \item{\dots}{
further arguments
}
}
\author{Thomas Roth \email{thomas.roth@tu-berlin.de}}

\note{
print.invisible is used to hide the internals of the method \code{\link{simProc}}
}
\seealso{
\code{\link{simProc}}
}
