\name{randomize}
\Rdversion{1.1}
\alias{randomize}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Randomization
}
\description{
function to do randomize the run order of factorial designs
}
\usage{
randomize(fdo, random.seed, so = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fdo}{
object of classs facDesign
}
  \item{random.seed}{
seed for randomness
}
  \item{so}{
TRUE/FALSE - use the standard Order?
}
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
