\name{rsmChoose}
\Rdversion{1.1}
\alias{rsmChoose}
\title{Choosing a response surface design from a table}
\description{
Designs displayed are central composite designs with orthogonal blocking and near rotatability. Choosing a design is done by clicking with the mouse into the appropriate field.
}
\usage{
rsmChoose()
}

\details{
Useful Central Composite Designs with orthogonal blocking and near rotatability
}
\value{
object of class facDesign
}
\references{
Response Surface Methodology (2009) Myers, Montgomery, Anderson-Cook p.329, table 7.5
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}

\seealso{
\code{\link{fracChoose}}
\code{\link{rsmDesign}}
}




