\name{[-methods}
\docType{methods}
\alias{[-methods}
\alias{[,ANY-method}
\alias{[,distrCollection-method}
\alias{[,facDesign-method}
\alias{[<-,facDesign-method}
\alias{[,gageRR-method}
\alias{[,nonStructure-method}
\title{Methods for Function [ in Package `base'}
\description{Methods for function \code{[} in Package `base'}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
ANY
}

\item{\code{signature(x = "distrCollection")}}{
object of class distrCollection
}

\item{\code{signature(x = "facDesign")}}{
object of class facDesign
}

\item{\code{signature(x = "gageRR")}}{
object fo class gageRR
}

\item{\code{signature(x = "nonStructure")}}{
nonStructure
}
}}
\keyword{methods}

