\name{standOrd-methods}
\docType{methods}
\alias{standOrd}
\alias{standOrd<-}
\alias{standOrd,facDesign-method}
\alias{standOrd<-,facDesign-method}
                        
\title{Get and set method}
\description{
Get and set the standard order of a factorial design.
}

\usage{
\S4method{standOrd}{facDesign}(x)
\S4method{standOrd}{facDesign}(x) <- value
}

\arguments{
\item{x}{a \code{\link{facDesign}} object}
\item{value}{new standOrd vector}
}



\section{Methods}{
\describe{

\item{\code{signature(object = "facDesign")}}{
So far used internally only.
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }

\seealso{
\url{http://www.r-qualitytools.org}
}
