\name{steepAscent-class}
\Rdversion{1.1}
\docType{class}
\alias{steepAscent-class}

\title{Class "steepAscent"}
\description{
steepAscent Class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("steepAscent", ...)}.
}

\section{Slots}{
  \describe{
    \item{\code{name}:}{Object of class \code{"character"} ~~ }
    \item{\code{X}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{response}:}{Object of class \code{"data.frame"} ~~ }
  }
}                                                                             
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\examples{
showClass("steepAscent")
}
\keyword{classes}
