\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,distrCollection-method}
\alias{summary,facDesign-method}
\alias{summary,gageRR-method}
\title{Methods for Function summary in Package `base'}
\description{Methods for function \code{summary} in Package `base'}
\section{Methods}{
\describe{

\item{\code{signature(object = "ANY")}}{
ANY
}

\item{\code{signature(object = "distrCollection")}}{
object of class distrCollection
}

\item{\code{signature(object = "facDesign")}}{
object of class facDesign
}

\item{\code{signature(object = "gageRR")}}{
object of class gageRR
}
}}

\author{Thomas Roth \email{thomas.roth@tu-berlin.de}}

\keyword{methods}

