\name{summaryFits}
\alias{summaryFits}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fit Summary 
}
\description{
Function to give an overview of fitted linear models for objects of class facDesign
}
\usage{
summaryFits(fdo, lmFit = TRUE, curvTest = TRUE, origFit = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fdo}{
object of class facDesign
}
  \item{lmFit}{
TRUE/FALSE - take the fits from the object fdo
}
  \item{curvTest}{
TRUE/FALSE - do curvature tests
}
  \item{origFit}{
TRUE/FALSS - display the fits in original values (as opposed to coded values)
}
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\note{
This is BETA!
}


