\name{taguchiFactor-class}
\Rdversion{1.1}
\docType{class}
\alias{taguchiFactor-class}

\title{Class "taguchiFactor"}
\description{
taguchiFactor Class
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("taguchiFactor", ...)}.
}


\section{Slots}{
  \describe{
    \item{\code{values}:}{Object of class \code{"ANY"} ~~ }
    \item{\code{name}:}{Object of class \code{"character"} ~~ }
    \item{\code{unit}:}{Object of class \code{"character"} ~~ }
    \item{\code{type}:}{Object of class \code{"character"} ~~ }
  }
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\examples{
showClass("taguchiFactor")
}
\keyword{classes}
