\name{values-methods}
\docType{methods}
\alias{values}
\alias{values<-}
\alias{values-methods}
\alias{values,taguchiDesign-method}
\alias{values<-,taguchiDesign-method}
\title{Get and Set methods}
\description{
Methods for function \code{values}
}

\usage{
\S4method{values}{taguchiDesign}(object)
\S4method{values}{taguchiDesign}(object) <- value
}
\arguments{
\item{object}{a taguchiDesign object}
\item{value}{a list of values}
}


\section{Methods}{
\describe{

\item{\code{signature(object = "taguchiDesign")}}{
Get and set the \code{values} for the factors in an object of class taguchiDesign
}
}}
\keyword{methods}

