\name{whiskersPlot-methods}
\docType{methods}
\alias{whiskersPlot-methods}
\alias{whiskersPlot,gageRR-method}
\title{Methods for Function whiskersPlot}
\description{
Methods for function \code{whiskersPlot}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "gageRR")}}{
function to create a whiskers plot for an object of class \code{\link{gageRR}}.
}

}}
\seealso{
\url{http://www.r-qualitytools.org}
}
\keyword{methods}
\keyword{design}
