\name{as.data.frame-methods}
\docType{methods}
\alias{as.data.frame-methods}
\alias{as.data.frame,ANY-method}
\alias{as.data.frame,desOpt-method}
\alias{as.data.frame,facDesign-method}
\alias{as.data.frame,mixDesign-method}
\alias{as.data.frame,taguchiDesign-method}
\alias{as.data.frame,pbDesign-method}
\alias{as.data.frame,steepAscent-method}
\alias{as.data.frame,gageRR-method}
\alias{as.data.frame,MSALinearity-method}
\title{Methods for Function as.data.frame in Package base}
\description{Methods for function \code{\link{as.data.frame}} in Package \code{base}}

\section{Methods}{
\describe{

\item{\code{signature(x = "MSALinearity")}}{
object of class \code{MSALinearity}}

\item{\code{signature(x = "ANY")}}{
ANY
}

\item{\code{signature(x = "desOpt")}}{
object of class \code{desOpt}
}

\item{\code{signature(x = "facDesign")}}{
object of class \code{\link{facDesign}}
}

\item{\code{signature(x = "mixDesign")}}{
object of class \code{\link{mixDesign}}
}

\item{\code{signature(x = "taguchiDesign")}}{
object of class \code{\link{taguchiDesign}}
}

\item{\code{signature(x = "pbDesign")}}{
object of class \code{\link{taguchiDesign}}
}

\item{\code{signature(x = "steepAscent")}}{
object of class \code{\link{steepAscent}}
}

\item{\code{signature(x = "gageRR")}}{
object of class \code{\link{gageRR}}
}}}

\keyword{methods}

\seealso{
\url{http://www.r-qualitytools.org}
}
