\name{code2real}
\alias{code2real}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coding
}
\description{
Function to calculate the real value of a coded value.
}
\usage{
code2real(low, high, codedValue)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{low}{
 numeric value giving the lower boundary.
}
  \item{high}{
 numeric value giving the higher boundary.
}
  \item{codedValue}{
 numeric avlue giving the coded value that will be calculated.
}
}
\author{Thomas Roth \email{thomas.roth@tu-berlin.de}}
\examples{
code2real(160, 200, 0)
}
\seealso{
\url{http://www.user.tu-berlin.de/kalicete/qualityTools}
}
