\name{confounds}
\alias{confounds}
\keyword{Design of Experiments}
\keyword{Six Sigma}
\title{
Confounded Effects
}
\description{
Function to display confounded effects of a fractional factorial design in a human readable way.
}
\usage{
confounds(x, depth = 2)
}
\arguments{
  \item{x}{
 needs to be an object of class \code{\link{facDesign}}.
}
  \item{depth}{
 numeric value - up to depth-way confounded interactions are printed
}
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\seealso{
\url{http://www.r-qualitytools.org}
}

