\name{effectPlot-methods}
\docType{methods}
\alias{effectPlot-methods}
\alias{effectPlot,facDesign-method}
\alias{effectPlot,taguchiDesign-method}
\title{Methods for Function effectPlot}
\description{
Methods for function \code{effectPlot}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "facDesign")}}{
function to create an effect plot for an object of class facDesign
}

\item{\code{signature(object = "taguchiDesign")}}{
function to create an effect plot for an object of class taguchiDesign
}
}}
\keyword{methods}
\keyword{design}
