\name{fracChoose}
\Rdversion{1.1}
\alias{fracChoose}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Choosing a fractional or full factorial design from a table.
}
\description{
Designs displayed are the classic minimum abberation designs. Choosing a design is done by clicking with the mouse into the appropriate field.
}
\usage{
fracChoose()
}

\value{
\code{fracChoose()} returns an object of class \code{\link{facDesign}}.
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\note{
 For an example in context which shows the usage of the function \code{fracChoose()}  
 please read the vignette for the package \code{\link{qualityTools}} at \url{http://www.r-qualitytools.org/Improve.html}.
}

\seealso{
\code{\link{facDesign}} \cr
\code{\link{fracDesign}}\cr
\code{\link{rsmChoose}} \cr
\code{\link{rsmDesign}} \cr
\url{http://www.r-qualitytools.org/Improve.html} \cr
Statistics for Experimenters (2005) Box, Hunter, Hunter p.272, table 6.22
}
