\name{rsmChoose}
\Rdversion{1.1}
\alias{rsmChoose}
\title{Choosing a response surface design from a table}
\description{
Designs displayed are central composite designs with orthogonal blocking and near rotatability. Choosing a design is done by clicking with the mouse into the appropriate field.
}
\usage{
rsmChoose()
}

\details{
Useful Central Composite Designs with orthogonal blocking and near rotatability
}
\value{
object of class \code{\link{facDesign}}.
}
\references{
MYERS, R.H.; MONTGOMERY, D.C.; ANDERSON-COOK, C.M.: Response Surface Methodology p.329, table 7.5. New Jersey: Wiley,2009.
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\note{
for an example for usage please visit \url{http://www.user.tu-berlin.de/kalicete/qualityTools/S5.html#SS6}.
}
\seealso{
\code{\link{fracChoose}}\cr
\code{\link{rsmDesign}}\cr
\url{http://www.user.tu-berlin.de/kalicete/qualityTools/S5.html#SS6}
}




