\name{rsmChoose}
\keyword{Design of Experiments}
\Rdversion{1.1}
\alias{rsmChoose}
\title{Choosing a response surface design from a table}
\description{
Designs displayed are central composite designs with orthogonal blocking and near rotatability. Choosing a design is done by clicking with the mouse into the appropriate field.
}
\usage{
rsmChoose()
}

\details{
Useful Central Composite Designs with orthogonal blocking and near rotatability
}
\value{
object of class \code{\link{facDesign}}.
}
\references{
\itemize{
 \item BOX, George. E. P. and HUNTER, J.S.: "Multifactor Experimental Designs for Exploring Response Surfaces",\cr
       The Annals of Mathematical Statistics 28, p.195-233, 1957 \cr
  \item MYERS, R.H.; MONTGOMERY, D.C.; ANDERSON-COOK, C.M.:\cr
       Response Surface Methodology (2009) p.329, table 7.5. New Jersey: Wiley.\cr
}


 
}
\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}
}
\note{
for an example for usage please visit \url{http://www.r-qualitytools.org/Improve.html}.
}
\seealso{
\code{\link{fracChoose}}\cr
\code{\link{rsmDesign}}\cr
\url{http://www.r-qualitytools.org/Improve.html}\cr
\url{http://webapps.r-qualitytools.org/brew/facDesign/rsmDesign.html}
}




