\name{[-methods}
\docType{methods}
\alias{[-methods}
\alias{[,ANY-method}
\alias{[,distrCollection-method}
\alias{[,facDesign-method}
\alias{[,steepAscent-method}
\alias{[<-,facDesign-method}
\alias{[,gageRR-method}
\alias{[,nonStructure-method}
\title{Methods for Function [ in Package base}
\description{Methods for function \code{[} in Package \code{base}.}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
ANY
}

\item{\code{signature(x = "distrCollection")}}{
object of class \code{distrCollection}.
}

\item{\code{signature(x = "facDesign")}}{
object of class \code{\link{facDesign}}.
}

\item{\code{signature(x = "steepAscent")}}{
object of class \code{\link{steepAscent}}.
}

\item{\code{signature(x = "gageRR")}}{
object fo class \code{\link{gageRR}}.
}

\item{\code{signature(x = "nonStructure")}}{
nonStructure
}
}}
\keyword{methods}

\seealso{
\url{http://www.r-qualitytools.org}
}