\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,distrCollection-method}
\alias{summary,facDesign-method}
\alias{summary,mixDesign-method}
\alias{summary,taguchiDesign-method}
\alias{summary,pbDesign-method}
\alias{summary,gageRR-method}
\alias{summary,MSALinearity-method}
\title{Methods for Function summary in Package base}
\description{Methods for function \code{summary} in Package \code{base}.}
\section{Methods}{
\describe{

\item{\code{signature(object = "ANY")}}{
ANY
}

\item{\code{signature(object = "MSALinearity")}}{
object of class \code{MSALinearity}.
}

\item{\code{signature(object = "distrCollection")}}{
object of class \code{distrCollection}.
}

\item{\code{signature(object = "facDesign")}}{
object of class \code{\link{facDesign}}.
}

\item{\code{signature(object = "mixDesign")}}{
object of class \code{\link{mixDesign}}.
}

\item{\code{signature(object = "taguchiDesign")}}{
object of class \code{\link{taguchiDesign}}.
}

\item{\code{signature(object = "pbDesign")}}{
object of class \code{\link{pbDesign}}.
}

\item{\code{signature(object = "gageRR")}}{
object of class \code{\link{gageRR}}.
}
}}

\author{
Thomas Roth \email{thomas.roth@tu-berlin.de}\cr
Etienne Stockhausen \email{stocdarf@mailbox.tu-berlin.de} 
}

\keyword{methods}
\seealso{
\url{http://www.r-qualitytools.org}
}
